feat(xo6/backup-job-configuration): create job scheduled composable
