+++ @xen-orchestra/web/src/components/backups/jobs/BackupJobsTable.vue
@@ -97,11 +97,9 @@
 </template>
 
 <script setup lang="ts">
+import { useXoBackupJobSchedules } from '@/composables/xo-backup-job-schedules.composable'
 import { useXoBackupUtils } from '@/composables/xo-backup-utils.composable.ts'
 import type { XoBackupJob } from '@/remote-resources/use-xo-backup-job-collection.ts'
-import { useXoBackupLogCollection } from '@/remote-resources/use-xo-backup-log-collection.ts'
-import { useXoScheduleCollection } from '@/remote-resources/use-xo-schedule-collection.ts'
-import type { XoBackupLog } from '@/types/xo/backup-log.type.ts'
 import type { IconName } from '@core/icons'
 import VtsDataTable from '@core/components/data-table/VtsDataTable.vue'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
@@ -119,7 +117,6 @@ import { usePagination } from '@core/composables/pagination.composable.ts'
 import { useRouteQuery } from '@core/composables/route-query.composable.ts'
 import { useTable } from '@core/composables/table.composable.ts'
 import { vTooltip } from '@core/directives/tooltip.directive.ts'
-import { createMapper } from '@core/packages/mapper'
 import { noop } from '@vueuse/shared'
 import { computed, ref } from 'vue'
 import { useI18n } from 'vue-i18n'
@@ -131,10 +128,8 @@ const { backupJobs } = defineProps<{
 
 const { t, d } = useI18n()
 
-const { schedulesByJobId } = useXoScheduleCollection()
-const { getLastNBackupLogsByJobId } = useXoBackupLogCollection()
 const { getModeLabels } = useXoBackupUtils()
-
+const { getLastThreeRunsStatuses, getTotalSchedules } = useXoBackupJobSchedules()
 const selectedBackupJobId = useRouteQuery('id')
 
 const searchQuery = ref('')
@@ -150,27 +145,6 @@ const filteredBackupJobs = computed(() => {
     Object.values(backupJob).some(value => String(value).toLocaleLowerCase().includes(searchTerm))
   )
 })
-
-const getRunStatusIcon = createMapper<XoBackupLog['status'], IconName>(
-  {
-    success: 'legacy:status:success',
-    skipped: 'legacy:status:warning',
-    interrupted: 'legacy:status:danger',
-    failure: 'legacy:status:danger',
-    pending: 'legacy:status:info',
-  },
-  'failure'
-)
-
-const getRunInfo = (backupLog: XoBackupLog, index: number) => ({
-  icon: getRunStatusIcon(backupLog.status),
-  tooltip: `${t('last-run-number', { n: index + 1 })}: ${d(backupLog.end ?? backupLog.start, 'datetime_short')}, ${t(backupLog.status)}`,
-})
-
-const getLastThreeRunsStatuses = (backupJob: XoBackupJob) =>
-  getLastNBackupLogsByJobId(backupJob.id).map((backupLog, index) => getRunInfo(backupLog, index))
-
-const getTotalSchedules = (backupJob: XoBackupJob) => schedulesByJobId.value.get(backupJob.id)?.length ?? 0
 
 const { visibleColumns, rows } = useTable('backup-jobs', filteredBackupJobs, {
   rowId: record => record.id,
